Consolidate logging config types: move LoggingConfig to utils, delete LoggingSetupConfig, slim OutputOptions
- Move LoggingConfig from config/data_types.py to utils/logging.py to
  eliminate the utils->config layer violation that LoggingSetupConfig
  was working around
- Delete LoggingSetupConfig bridge type; setup_logging now takes
  LoggingConfig directly plus a default_host_dir parameter
- Remove logging-related fields from OutputOptions (console_level,
  log_level, log_file_path, is_log_commands, is_log_command_output,
  is_log_env_vars) -- these are now on the resolved LoggingConfig
- Add log_level and log_file_path fields to LoggingConfig to carry
  CLI overrides alongside TOML defaults
- Add is_quiet field to OutputOptions for the --quiet output guard
- Store resolved LoggingConfig on ctx.meta["logging_config"] so
  callers like create.py can access logging levels for LoggingSuppressor
- Update parse_output_options to return (OutputOptions, LoggingConfig)
- Update all test files to use LoggingConfig directly
