+++ libs/mng/imbue/mng/utils/logging_test.py
@@ -9,14 +9,12 @@ from loguru import logger
 from imbue.imbue_common.logging import _format_arg_value
 from imbue.imbue_common.logging import log_call
 from imbue.mng.config.data_types import MngContext
-from imbue.mng.config.data_types import OutputOptions
 from imbue.mng.primitives import LogLevel
-from imbue.mng.primitives import OutputFormat
 from imbue.mng.utils.logging import BUILD_COLOR
 from imbue.mng.utils.logging import BufferedMessage
 from imbue.mng.utils.logging import DEBUG_COLOR
 from imbue.mng.utils.logging import ERROR_COLOR
-from imbue.mng.utils.logging import LoggingSetupConfig
+from imbue.mng.utils.logging import LoggingConfig
 from imbue.mng.utils.logging import LoggingSuppressor
 from imbue.mng.utils.logging import RESET_COLOR
 from imbue.mng.utils.logging import WARNING_COLOR
@@ -26,20 +24,6 @@ from imbue.mng.utils.logging import _resolve_log_dir
 from imbue.mng.utils.logging import _rotate_old_logs
 from imbue.mng.utils.logging import remove_console_handlers
 from imbue.mng.utils.logging import setup_logging
-
-
-def _make_logging_config(output_opts: OutputOptions, mng_ctx: MngContext) -> LoggingSetupConfig:
-    """Build a LoggingSetupConfig from test fixtures."""
-    return LoggingSetupConfig(
-        console_level=output_opts.console_level,
-        log_level=output_opts.log_level,
-        log_file_path=output_opts.log_file_path,
-        file_level=mng_ctx.config.logging.file_level,
-        max_log_size_mb=mng_ctx.config.logging.max_log_size_mb,
-        max_log_files=mng_ctx.config.logging.max_log_files,
-        log_dir=mng_ctx.config.logging.log_dir,
-        default_host_dir=mng_ctx.config.default_host_dir,
-    )
 
 
 def test_resolve_log_dir_uses_absolute_path(mng_test_prefix: str) -> None:
@@ -100,14 +84,9 @@ def test_setup_logging_creates_log_dir(temp_mng_ctx: MngContext) -> None:
     log_dir = temp_mng_ctx.config.default_host_dir / temp_mng_ctx.config.logging.log_dir
     assert not log_dir.exists()
 
-    output_opts = OutputOptions(
-        output_format=OutputFormat.HUMAN,
-        console_level=LogLevel.INFO,
-        is_log_commands=True,
-        is_log_command_output=False,
-    )
+    logging_config = LoggingConfig(console_level=LogLevel.INFO)
 
-    setup_logging(_make_logging_config(output_opts, temp_mng_ctx))
+    setup_logging(logging_config, default_host_dir=temp_mng_ctx.config.default_host_dir)
 
     assert log_dir.exists()
     assert log_dir.is_dir()
@@ -116,14 +95,9 @@ def test_setup_logging_creates_log_dir(temp_mng_ctx: MngContext) -> None:
 def test_setup_logging_creates_log_file(temp_mng_ctx: MngContext) -> None:
     """setup_logging should create a log file."""
     log_dir = temp_mng_ctx.config.default_host_dir / temp_mng_ctx.config.logging.log_dir
-    output_opts = OutputOptions(
-        output_format=OutputFormat.HUMAN,
-        console_level=LogLevel.INFO,
-        is_log_commands=True,
-        is_log_command_output=False,
-    )
+    logging_config = LoggingConfig(console_level=LogLevel.INFO)
 
-    setup_logging(_make_logging_config(output_opts, temp_mng_ctx))
+    setup_logging(logging_config, default_host_dir=temp_mng_ctx.config.default_host_dir)
 
     log_files = list(log_dir.glob("*.json"))
     assert len(log_files) >= 1
@@ -133,15 +107,12 @@ def test_setup_logging_uses_custom_log_file_path(tmp_path: Path, temp_mng_ctx: M
     """setup_logging should create log file at custom path when log_file_path is provided."""
     custom_log_path = tmp_path / "custom_log.json"
 
-    output_opts = OutputOptions(
-        output_format=OutputFormat.HUMAN,
+    logging_config = LoggingConfig(
         console_level=LogLevel.INFO,
         log_file_path=custom_log_path,
-        is_log_commands=True,
-        is_log_command_output=False,
     )
 
-    setup_logging(_make_logging_config(output_opts, temp_mng_ctx))
+    setup_logging(logging_config, default_host_dir=temp_mng_ctx.config.default_host_dir)
 
     assert custom_log_path.exists()
 
@@ -152,15 +123,12 @@ def test_setup_logging_creates_parent_dirs_for_custom_log_path(tmp_path: Path, t
 
     assert not custom_log_path.parent.exists()
 
-    output_opts = OutputOptions(
-        output_format=OutputFormat.HUMAN,
+    logging_config = LoggingConfig(
         console_level=LogLevel.INFO,
         log_file_path=custom_log_path,
-        is_log_commands=True,
-        is_log_command_output=False,
     )
 
-    setup_logging(_make_logging_config(output_opts, temp_mng_ctx))
+    setup_logging(logging_config, default_host_dir=temp_mng_ctx.config.default_host_dir)
 
     assert custom_log_path.parent.exists()
     assert custom_log_path.exists()
@@ -182,15 +150,12 @@ def test_setup_logging_expands_user_in_custom_log_path(tmp_path: Path, temp_mng_
     relative_path = log_subdir.relative_to(home_dir)
     tilde_path = Path("~") / relative_path / "expanded_log.json"
 
-    output_opts = OutputOptions(
-        output_format=OutputFormat.HUMAN,
+    logging_config = LoggingConfig(
         console_level=LogLevel.INFO,
         log_file_path=tilde_path,
-        is_log_commands=True,
-        is_log_command_output=False,
     )
 
-    setup_logging(_make_logging_config(output_opts, temp_mng_ctx))
+    setup_logging(logging_config, default_host_dir=temp_mng_ctx.config.default_host_dir)
 
     expanded_path = home_dir / relative_path / "expanded_log.json"
     assert expanded_path.exists()
@@ -338,13 +303,8 @@ def test_logging_suppressor_initial_state() -> None:
 
 def test_logging_suppressor_enable_sets_suppressed() -> None:
     """Enable should set suppressed state to True."""
-    output_opts = OutputOptions(
-        output_format=OutputFormat.HUMAN,
-        console_level=LogLevel.INFO,
-    )
-
     try:
-        LoggingSuppressor.enable(output_opts.console_level, output_opts.log_level)
+        LoggingSuppressor.enable(LogLevel.INFO, LogLevel.NONE)
         assert LoggingSuppressor.is_suppressed()
     finally:
         LoggingSuppressor.disable_and_replay(clear_screen=False)
@@ -352,12 +312,7 @@ def test_logging_suppressor_enable_sets_suppressed() -> None:
 
 def test_logging_suppressor_disable_clears_suppressed() -> None:
     """Disable should set suppressed state to False."""
-    output_opts = OutputOptions(
-        output_format=OutputFormat.HUMAN,
-        console_level=LogLevel.INFO,
-    )
-
-    LoggingSuppressor.enable(output_opts.console_level, output_opts.log_level)
+    LoggingSuppressor.enable(LogLevel.INFO, LogLevel.NONE)
     assert LoggingSuppressor.is_suppressed()
 
     LoggingSuppressor.disable_and_replay(clear_screen=False)
@@ -366,13 +321,8 @@ def test_logging_suppressor_disable_clears_suppressed() -> None:
 
 def test_logging_suppressor_buffers_messages() -> None:
     """Suppressor should buffer messages while suppression is enabled."""
-    output_opts = OutputOptions(
-        output_format=OutputFormat.HUMAN,
-        console_level=LogLevel.INFO,
-    )
-
     try:
-        LoggingSuppressor.enable(output_opts.console_level, output_opts.log_level)
+        LoggingSuppressor.enable(LogLevel.INFO, LogLevel.NONE)
 
         # Log some messages
         logger.info("Test message 1")
@@ -389,14 +339,9 @@ def test_logging_suppressor_buffers_messages() -> None:
 
 def test_logging_suppressor_respects_buffer_size() -> None:
     """Suppressor should limit buffer to specified size."""
-    output_opts = OutputOptions(
-        output_format=OutputFormat.HUMAN,
-        console_level=LogLevel.INFO,
-    )
-
     try:
         # Enable with small buffer
-        LoggingSuppressor.enable(output_opts.console_level, output_opts.log_level, buffer_size=3)
+        LoggingSuppressor.enable(LogLevel.INFO, LogLevel.NONE, buffer_size=3)
 
         # Log more messages than buffer size
         for i in range(10):
@@ -411,12 +356,7 @@ def test_logging_suppressor_respects_buffer_size() -> None:
 
 def test_logging_suppressor_clears_buffer_on_disable() -> None:
     """Suppressor should clear buffer after disable_and_replay."""
-    output_opts = OutputOptions(
-        output_format=OutputFormat.HUMAN,
-        console_level=LogLevel.INFO,
-    )
-
-    LoggingSuppressor.enable(output_opts.console_level, output_opts.log_level)
+    LoggingSuppressor.enable(LogLevel.INFO, LogLevel.NONE)
     logger.info("Test message")
     assert len(LoggingSuppressor.get_buffered_messages()) >= 1
 
@@ -426,18 +366,13 @@ def test_logging_suppressor_clears_buffer_on_disable() -> None:
 
 def test_logging_suppressor_enable_is_idempotent() -> None:
     """Calling enable twice should not reset buffer."""
-    output_opts = OutputOptions(
-        output_format=OutputFormat.HUMAN,
-        console_level=LogLevel.INFO,
-    )
-
     try:
-        LoggingSuppressor.enable(output_opts.console_level, output_opts.log_level)
+        LoggingSuppressor.enable(LogLevel.INFO, LogLevel.NONE)
         logger.info("First message")
         initial_count = len(LoggingSuppressor.get_buffered_messages())
 
         # Enable again (should be no-op)
-        LoggingSuppressor.enable(output_opts.console_level, output_opts.log_level)
+        LoggingSuppressor.enable(LogLevel.INFO, LogLevel.NONE)
         assert len(LoggingSuppressor.get_buffered_messages()) == initial_count
     finally:
         LoggingSuppressor.disable_and_replay(clear_screen=False)
@@ -445,12 +380,7 @@ def test_logging_suppressor_enable_is_idempotent() -> None:
 
 def test_logging_suppressor_disable_is_idempotent() -> None:
     """Calling disable_and_replay twice should be safe."""
-    output_opts = OutputOptions(
-        output_format=OutputFormat.HUMAN,
-        console_level=LogLevel.INFO,
-    )
-
-    LoggingSuppressor.enable(output_opts.console_level, output_opts.log_level)
+    LoggingSuppressor.enable(LogLevel.INFO, LogLevel.NONE)
     LoggingSuppressor.disable_and_replay(clear_screen=False)
 
     # Second disable should not error
@@ -474,13 +404,10 @@ def test_buffered_message_tracks_stderr_destination() -> None:
 
 def test_remove_console_handlers_clears_handler_ids(temp_mng_ctx: MngContext) -> None:
     """remove_console_handlers should clear _console_handler_ids dict."""
-    output_opts = OutputOptions(
-        output_format=OutputFormat.HUMAN,
-        console_level=LogLevel.INFO,
-    )
+    logging_config = LoggingConfig(console_level=LogLevel.INFO)
 
     # Setup logging to populate console handler IDs
-    setup_logging(_make_logging_config(output_opts, temp_mng_ctx))
+    setup_logging(logging_config, default_host_dir=temp_mng_ctx.config.default_host_dir)
     assert len(_console_handler_ids) > 0
 
     # Remove handlers
@@ -492,12 +419,9 @@ def test_remove_console_handlers_clears_handler_ids(temp_mng_ctx: MngContext) ->
 
 def test_remove_console_handlers_is_idempotent(temp_mng_ctx: MngContext) -> None:
     """Calling remove_console_handlers twice should not error."""
-    output_opts = OutputOptions(
-        output_format=OutputFormat.HUMAN,
-        console_level=LogLevel.INFO,
-    )
+    logging_config = LoggingConfig(console_level=LogLevel.INFO)
 
-    setup_logging(_make_logging_config(output_opts, temp_mng_ctx))
+    setup_logging(logging_config, default_host_dir=temp_mng_ctx.config.default_host_dir)
     remove_console_handlers()
 
     # Second call should not raise an error
@@ -532,13 +456,12 @@ def test_setup_logging_writes_to_current_stderr_after_stream_replacement(
 
     Both the user-facing handler and the diagnostic handler write to stderr.
     """
-    output_opts = OutputOptions(
-        output_format=OutputFormat.HUMAN,
+    logging_config = LoggingConfig(
         console_level=LogLevel.INFO,
         log_level=LogLevel.DEBUG,
     )
 
-    setup_logging(_make_logging_config(output_opts, temp_mng_ctx))
+    setup_logging(logging_config, default_host_dir=temp_mng_ctx.config.default_host_dir)
 
     # Replace sys.stderr with a StringIO to simulate pytest's capture mechanism
     original_stderr = sys.stderr
